Update login_window.py
I believe that the boot for the Login Window has been delayed by 0.5 seconds due to having to load images for both the window icon and the HMS logo as well as the hide/show password icon image. 

I have reduced code by minimising the the location imports so that importing the images from assets/images/img1.jpg is shortened too assets/img1.jpg but I have not seen much difference. perhaps that the window now has to load images is that it takes longer to boot. 

Other than that all is good!